Se movió la estructura de app a router y se implementó redux thunk para redux asíncrono
