+++ actions/stateCar.js
@@ -6,11 +6,6 @@ export const switchStateCar = ( state ) => ({
     payload: state
 });
 
-export const changeSpeedometer = ( velocity ) => ({
-    type: types.speedometer,
-    payload: velocity
-})
-
 export const changeStateDoors = ( state ) => ({
     type: types.doors,
     payload: state
