+++ components/Molecules/Odometer.js
@@ -1,4 +1,4 @@
-import { View, Text } from 'native-base'
+import { View, Text } from 'react-native'
 import React from 'react'
 
 export const Odometer = ({ kms }) => {
