+++ package.json
@@ -21,7 +21,6 @@
     "axios": "^0.21.1",
     "buffer": "^5.7.1",
     "install": "^0.13.0",
-    "native-base": "^2.15.2",
     "react": "17.0.1",
     "react-native": "0.64.0",
     "react-native-config": "^1.4.2",
